not_dot_net/frontend/admin_settings.py
"""Admin settings page — auto-generated forms from config registry."""

import json
import logging
from enum import Enum

from nicegui import ui
from pydantic import BaseModel, ValidationError
from yaml import safe_dump, safe_load

from not_dot_net.backend.app_config import get_registry
from not_dot_net.backend.audit import log_audit
from not_dot_net.backend.data_io import export_all, import_all
from not_dot_net.frontend.admin_roles import render as render_roles
from not_dot_net.frontend.i18n import t

logger = logging.getLogger(__name__)


def _is_enum(annotation) -> bool:
    return isinstance(annotation, type) and issubclass(annotation, Enum)


def _is_complex(schema: type[BaseModel]) -> bool:
    """Check if a schema has nested models or dicts — use YAML editor."""
    for field_info in schema.model_fields.values():
        annotation = field_info.annotation
        if annotation is dict or (hasattr(annotation, "__origin__") and annotation.__origin__ is dict):
            return True
        if isinstance(annotation, type) and issubclass(annotation, BaseModel):
            return True
    return False


async def render(user):
    """Render the admin settings tab content."""
    from not_dot_net.backend.permissions import check_permission
    await check_permission(user, "manage_settings")

    with ui.expansion("Roles", icon="admin_panel_settings").classes("w-full"):
        await render_roles(user)

    with ui.expansion(t("import_export"), icon="swap_vert").classes("w-full"):
        _render_import_export(user)

    registry = get_registry()

    for prefix, cfg_section in sorted(registry.items()):
        current = await cfg_section.get()
        schema = cfg_section.schema

        with ui.expansion(cfg_section.label, icon="settings").classes("w-full"):
            if _is_complex(schema):
                await _render_yaml_editor(prefix, cfg_section, current, user)
            else:
                await _render_form(prefix, cfg_section, current, user)


async def _render_form(prefix, cfg_section, current, user):
    """Auto-generate form fields from Pydantic model."""
    inputs = {}
    schema = cfg_section.schema
    data = current.model_dump()

    for field_name, field_info in schema.model_fields.items():
        annotation = field_info.annotation
        value = data.get(field_name, field_info.default)

        hint = field_info.description

        if annotation is bool:
            widget = ui.switch(field_name, value=value)
        elif _is_enum(annotation):
            options = {m.value: m.value for m in annotation}
            widget = ui.select(
                options, label=field_name, value=value,
            ).classes("w-full")
        elif annotation is int:
            widget = ui.number(field_name, value=value)
        elif annotation is str:
            widget = ui.input(field_name, value=value).classes("w-full")
        elif annotation == list[str]:
            widget = ui.input(
                field_name,
                value=", ".join(value) if isinstance(value, list) else str(value),
            ).classes("w-full")
            hint = hint or "Comma-separated values"
        else:
            widget = ui.input(field_name, value=str(value)).classes("w-full")

        if hint:
            widget.tooltip(hint)
        inputs[field_name] = widget

    async def save():
        update = {}
        for field_name, field_info in schema.model_fields.items():
            widget = inputs[field_name]
            annotation = field_info.annotation
            if annotation is bool:
                update[field_name] = widget.value
            elif annotation is int:
                update[field_name] = int(widget.value)
            elif annotation == list[str]:
                update[field_name] = [s.strip() for s in widget.value.split(",") if s.strip()]
            else:
                update[field_name] = widget.value
        try:
            new_config = schema.model_validate(update)
            await cfg_section.set(new_config)
            await log_audit("settings", "update", actor_id=user.id, actor_email=user.email, detail=f"section={prefix}")
            ui.notify(t("settings_saved"), color="positive")
        except ValidationError as e:
            ui.notify(str(e), color="negative")

    async def reset():
        await cfg_section.reset()
        await log_audit("settings", "reset", actor_id=user.id, actor_email=user.email, detail=f"section={prefix}")
        ui.notify(t("settings_reset"), color="info")

    with ui.row():
        ui.button(t("save"), on_click=save).props("color=primary")
        ui.button(t("reset_defaults"), on_click=reset).props("flat color=grey")


async def _render_yaml_editor(prefix, cfg_section, current, user):
    """YAML code editor for complex config sections."""
    yaml_str = safe_dump(current.model_dump(), default_flow_style=False, allow_unicode=True)
    editor = ui.codemirror(yaml_str, language="yaml").classes("w-full").style("min-height: 300px")

    async def save():
        try:
            data = safe_load(editor.value)
            new_config = cfg_section.schema.model_validate(data)
            await cfg_section.set(new_config)
            await log_audit("settings", "update", actor_id=user.id, actor_email=user.email, detail=f"section={prefix}")
            ui.notify(t("settings_saved"), color="positive")
        except Exception as e:
            ui.notify(str(e), color="negative")

    async def reset():
        await cfg_section.reset()
        default = cfg_section.schema()
        editor.value = safe_dump(default.model_dump(), default_flow_style=False, allow_unicode=True)
        await log_audit("settings", "reset", actor_id=user.id, actor_email=user.email, detail=f"section={prefix}")
        ui.notify(t("settings_reset"), color="info")

    with ui.row():
        ui.button(t("save"), on_click=save).props("color=primary")
        ui.button(t("reset_defaults"), on_click=reset).props("flat color=grey")


def _render_import_export(user):
    ui.label(t("import_export_help")).classes("text-sm text-grey mb-2")

    with ui.row().classes("items-center gap-2"):
        async def do_export():
            data = await export_all()
            content = json.dumps(data, indent=2, ensure_ascii=False)
            ui.download(content.encode(), "not-dot-net-export.json")
            await log_audit(
                "settings", "export",
                actor_id=user.id, actor_email=user.email,
                detail=f"pages={len(data.get('pages', []))} resources={len(data.get('resources', []))}",
            )

        ui.button(t("export_all"), icon="download", on_click=do_export).props("color=primary")

    ui.separator().classes("my-3")

    replace_toggle = ui.switch(t("import_replace")).tooltip(t("import_replace_help"))

    async def handle_upload(e):
        await _handle_import_upload(e, replace=replace_toggle.value, user=user)

    ui.upload(
        label=t("import_file"),
        on_upload=handle_upload,
        auto_upload=True,
    ).props("accept=.json").classes("w-full max-w-md")


async def _handle_import_upload(e, *, replace: bool, user):
    try:
        data = await e.file.json()
    except (json.JSONDecodeError, UnicodeDecodeError, ValueError) as exc:
        logger.warning("Import: invalid JSON: %s", exc)
        ui.notify(t("import_invalid_json"), color="negative")
        return
    try:
        result = await import_all(data, replace=replace)
    except Exception:
        logger.exception("Import failed")
        ui.notify(t("import_failed"), color="negative")
        return
    if not result:
        ui.notify(t("import_nothing"), color="warning")
        return
    ui.notify("; ".join(
        f"{entity}: {c['created']} created, {c['updated']} updated, {c['skipped']} skipped"
        for entity, c in result.items()
    ), color="positive", multi_line=True)
    await log_audit(
        "settings", "import",
        actor_id=user.id, actor_email=user.email,
        detail=json.dumps(result),
    )
